L4G-5 Launch activity for search preferences, small refactor,
