+++ Client/app/src/main/AndroidManifest.xml
@@ -15,20 +15,24 @@
 
         <activity
             android:name=".MainViewActivity"
-            android:label="@string/title_activity_register_view"
+            android:label="@string/title_activity_main_view"
             android:parentActivityName=".ui.login.LoginActivity"
             android:theme="@style/Theme.Client.NoActionBar" />
         <activity
+            android:name=".ui.preferenceSearch.SearchPreferenceActivity"
+            android:label="@string/title_activity_search_profile_view"
+            android:parentActivityName=".MainViewActivity"
+            android:theme="@style/Theme.Client.NoActionBar"/>
+        <activity
             android:name=".ui.registration.RegistrationActivity"
-            android:label="@string/title_activity_main_view"
+            android:label="@string/title_activity_register_view"
             android:parentActivityName=".ui.login.LoginActivity"
-            android:theme="@style/Theme.Client.NoActionBar"></activity>
+            android:theme="@style/Theme.Client.NoActionBar" />
         <activity
             android:name=".ui.login.LoginActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
