+++ Client/app/src/main/java/com/Client/Table/MainViewActivity.kt
@@ -3,7 +3,9 @@ package com.Client.Table
 import android.content.Intent
 import android.os.Bundle
 import android.view.Menu
+import android.view.MenuItem
 import android.widget.Button
+import android.widget.Toast
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.android.material.snackbar.Snackbar
 import com.google.android.material.navigation.NavigationView
@@ -15,11 +17,12 @@ import androidx.navigation.ui.setupWithNavController
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
+import com.Client.Table.ui.preferenceSearch.SearchPreferenceActivity
+import com.Client.Table.ui.registration.RegistrationActivity
 
 class MainViewActivity : AppCompatActivity() {
 
     private lateinit var appBarConfiguration: AppBarConfiguration
-    lateinit var settings_btn : Button
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -50,11 +53,22 @@ class MainViewActivity : AppCompatActivity() {
     override fun onBackPressed() {
         finish()
     }
+
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
         // Inflate the menu; this adds items to the action bar if it is present.
         menuInflater.inflate(R.menu.main_view, menu)
+        val search_pref_item: MenuItem = menu.findItem(R.id.action_settings)
+            search_pref_item.setOnMenuItemClickListener {
+                //start intend
+                val intent = Intent(this, SearchPreferenceActivity::class.java).apply {  }
+                //add variable with observer, set observed variable with return value of Activity for changing search view?
+                startActivity(intent)
+                true
+            }
+
         return true
     }
+
 
     override fun onSupportNavigateUp(): Boolean {
         val navController = findNavController(R.id.nav_host_fragment)
