+++ Client/app/src/main/res/menu/main_view.xml
@@ -4,6 +4,6 @@
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
-        android:title="@string/action_settings"
+        android:title="@string/search_profile_preference"
         app:showAsAction="never" />
 </menu>
