+++ Client/app/src/main/res/values/strings.xml
@@ -14,6 +14,7 @@
     <string name="login_text">"Login to Looking 4 Group"</string>
     <string name="title_activity_main_view">MainViewActivity</string>
     <string name="title_activity_register_view">RegisterViewActivity</string>
+    <string name="title_activity_search_profile_view">SearchProfileView</string>
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
     <string name="nav_header_title">Android Studio</string>
@@ -67,6 +68,7 @@
         <item>70.0</item>
     </array>
     <string name="crossed_ttrpg"> <strike> T</strike> </string>
+    <string name="search_profile_preference"> search preferences </string>
